requirements.txt
Flask
Flask-RESTful
gunicorn
itsdangerous
Jinja2
peewee
Werkzeug
tensorflow-cpu==2.6.0
numpy==1.19.5
